+++ x/uibc/quota/msg_server.go
@@ -1,4 +1,4 @@
-package keeper
+package quota
 
 import (
 	"context"
@@ -11,11 +11,11 @@ import (
 var _ uibc.MsgServer = msgServer{}
 
 type msgServer struct {
-	kb Builder
+	kb KeeperBuilder
 }
 
 // NewMsgServerImpl returns an implementation of uibc.MsgServer
-func NewMsgServerImpl(kb Builder) uibc.MsgServer {
+func NewMsgServerImpl(kb KeeperBuilder) uibc.MsgServer {
 	return &msgServer{kb: kb}
 }
 
@@ -34,7 +34,7 @@ func (m msgServer) GovUpdateQuota(ctx context.Context, msg *uibc.MsgGovUpdateQuo
 		return nil, err
 	}
 
-	if err := k.UpdateQuotaParams(msg.Total, msg.PerDenom, msg.QuotaDuration, byEmergencyGroup); err != nil {
+	if err := k.UpdateQuotaParams(msg, byEmergencyGroup); err != nil {
 		return nil, err
 	}
 	return &uibc.MsgGovUpdateQuotaResponse{}, nil