+++ x/uibc/quota/quota.go
@@ -1,4 +1,4 @@
-package keeper
+package quota
 
 import (
 	"strings"
@@ -6,7 +6,7 @@ import (
 
 	sdkmath "cosmossdk.io/math"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	transfertypes "github.com/cosmos/ibc-go/v7/modules/apps/transfer/types"
+	ics20types "github.com/cosmos/ibc-go/v7/modules/apps/transfer/types"
 	channeltypes "github.com/cosmos/ibc-go/v7/modules/core/04-channel/types"
 	"github.com/cosmos/ibc-go/v7/modules/core/exported"
 
@@ -28,7 +28,7 @@ func (k Keeper) GetAllOutflows() (sdk.DecCoins, error) {
 // GetTokenOutflows returns sum of denom outflows in USD value in the DecCoin structure.
 func (k Keeper) GetTokenOutflows(denom string) sdk.DecCoin {
 	// When token outflow is not stored in store it will return 0
-	amount, _ := store.GetDec(k.store, keyTotalOutflow(denom), "total_outflow")
+	amount, _ := store.GetDec(k.store, keyTokenOutflow(denom), "total_outflow")
 	return sdk.NewDecCoinFromDec(denom, amount)
 }
 
@@ -40,22 +40,22 @@ func (k Keeper) SetTokenOutflows(outflows sdk.DecCoins) {
 	}
 }
 
-// SetTotalOutflowSum save the total outflow of ibc-transfer amount.
-func (k Keeper) SetTotalOutflowSum(amount sdk.Dec) {
-	err := store.SetDec(k.store, keyTotalOutflows, amount, "total_outflow_sum")
+// SetOutflowSum save the total outflow of ibc-transfer amount.
+func (k Keeper) SetOutflowSum(amount sdk.Dec) {
+	err := store.SetDec(k.store, keyOutflowSum, amount, "total_outflow_sum")
 	util.Panic(err)
 }
 
-// GetTotalOutflow returns the total outflow of ibc-transfer amount.
-func (k Keeper) GetTotalOutflow() sdk.Dec {
+// GetOutflowSum returns the total outflow of ibc-transfer amount.
+func (k Keeper) GetOutflowSum() sdk.Dec {
 	// When total outflow is not stored in store it will return 0
-	amount, _ := store.GetDec(k.store, keyTotalOutflows, "total_outflow")
+	amount, _ := store.GetDec(k.store, keyOutflowSum, "total_outflow")
 	return amount
 }
 
 // SetTokenOutflow save the outflows of denom into store.
 func (k Keeper) SetTokenOutflow(outflow sdk.DecCoin) {
-	key := keyTotalOutflow(outflow.Denom)
+	key := keyTokenOutflow(outflow.Denom)
 	err := store.SetDec(k.store, key, outflow.Amount, "total_outflow")
 	util.Panic(err)
 }
@@ -89,16 +89,16 @@ func (k Keeper) GetTokenInflow(denom string) sdk.DecCoin {
 	return sdk.NewDecCoinFromDec(denom, amount)
 }
 
-// GetTotalInflow returns the total inflow of ibc-transfer amount.
-func (k Keeper) GetTotalInflow() sdk.Dec {
+// GetInflowSum returns the total inflow of ibc-transfer amount.
+func (k Keeper) GetInflowSum() sdk.Dec {
 	// When total inflow is not stored in store it will return 0
-	amount, _ := store.GetDec(k.store, keyTotalInflows, "total_inflows")
+	amount, _ := store.GetDec(k.store, keyInflowSum, "total_inflows")
 	return amount
 }
 
-// SetTotalInflow save the total inflow of ibc-transfer amount.
-func (k Keeper) SetTotalInflow(amount sdk.Dec) {
-	err := store.SetDec(k.store, keyTotalInflows, amount, "total_inflows")
+// SetInflowSum save the total inflow of ibc-transfer amount.
+func (k Keeper) SetInflowSum(amount sdk.Dec) {
+	err := store.SetDec(k.store, keyInflowSum, amount, "total_inflows")
 	util.Panic(err)
 }
 
@@ -121,12 +121,12 @@ func (k Keeper) ResetAllQuotas() error {
 	}
 	zero := sdk.NewDec(0)
 	// outflows
-	k.SetTotalOutflowSum(zero)
+	k.SetOutflowSum(zero)
 	ps := k.PrefixStore(keyPrefixDenomOutflows)
 	store.DeleteByPrefixStore(ps)
 
 	// inflows
-	k.SetTotalInflow(zero)
+	k.SetInflowSum(zero)
 	ps = k.PrefixStore(keyPrefixDenomInflows)
 	store.DeleteByPrefixStore(ps)
 	return nil
@@ -150,25 +150,24 @@ func (k Keeper) CheckAndUpdateQuota(denom string, newOutflow sdkmath.Int) error 
 	inToken := k.GetTokenInflow(denom)
 	if !params.TokenQuota.IsZero() {
 		if o.Amount.GT(params.TokenQuota) ||
-			o.Amount.GT(params.InflowOutflowQuotaTokenBase.Add((params.InflowOutflowQuotaRate.Mul(inToken.Amount)))) {
+			o.Amount.GT(params.InflowOutflowTokenQuotaBase.Add((params.InflowOutflowQuotaRate.Mul(inToken.Amount)))) {
 			return uibc.ErrQuotaExceeded
 		}
 	}
 
 	// Allow outflow either of two conditions
-	// 1. Total Outflow Sum <= Total Outflow Quota
-	// or
-	// 2. Total Outflow Sum <= params.InflowOutflowQuotaBase + (params.InflowOutflowQuotaRate * sum of all inflows)
-	totalOutflowSum := k.GetTotalOutflow().Add(exchangePrice)
-	ttlInSum := k.GetTotalInflow()
+	// 1. Outflow Sum <= Total Outflow Quota
+	// 2. OR Outflow Sum <= params.InflowOutflowQuotaBase + (params.InflowOutflowQuotaRate * sum of all inflows)
+	outflowSum := k.GetOutflowSum().Add(exchangePrice)
+	inflowSum := k.GetInflowSum()
 	if !params.TotalQuota.IsZero() {
-		if totalOutflowSum.GT(params.TotalQuota) ||
-			totalOutflowSum.GT(params.InflowOutflowQuotaBase.Add(ttlInSum.Mul(params.InflowOutflowQuotaRate))) {
+		if outflowSum.GT(params.TotalQuota) ||
+			outflowSum.GT(params.InflowOutflowQuotaBase.Add(inflowSum.Mul(params.InflowOutflowQuotaRate))) {
 			return uibc.ErrQuotaExceeded
 		}
 	}
 	k.SetTokenOutflow(o)
-	k.SetTotalOutflowSum(totalOutflowSum)
+	k.SetOutflowSum(outflowSum)
 
 	return nil
 }
@@ -224,30 +223,31 @@ func (k Keeper) UndoUpdateQuota(denom string, amount sdkmath.Int) error {
 	}
 	k.SetTokenOutflow(o)
 
-	totalOutflowSum := k.GetTotalOutflow()
-	k.SetTotalOutflowSum(totalOutflowSum.Sub(exchangePrice))
+	outflowSum := k.GetOutflowSum()
+	k.SetOutflowSum(outflowSum.Sub(exchangePrice))
 	return nil
 }
 
 // RecordIBCInflow will save the inflow amount if token is registered otherwise it will skip
-func (k Keeper) RecordIBCInflow(ctx sdk.Context,
-	packet channeltypes.Packet, dataDenom, dataAmount string, isSourceChain bool,
+func (k Keeper) RecordIBCInflow(packet channeltypes.Packet, denom, amount string,
 ) exported.Acknowledgement {
+
 	// if chain is recevier and sender chain is source then we need create ibc_denom (ibc/hash(channel,denom)) to
 	// check ibc_denom is exists in leverage token registry
-	if isSourceChain {
+	if !ics20types.SenderChainIsSource(packet.GetSourcePort(), packet.GetSourceChannel(), denom) {
 		// since SendPacket did not prefix the denomination, we must prefix denomination here
-		sourcePrefix := transfertypes.GetDenomPrefix(packet.GetDestPort(), packet.GetDestChannel())
+		sourcePrefix := ics20types.GetDenomPrefix(packet.GetDestPort(), packet.GetDestChannel())
 		// NOTE: sourcePrefix contains the trailing "/"
-		prefixedDenom := sourcePrefix + dataDenom
+		prefixedDenom := sourcePrefix + denom
 		// construct the denomination trace from the full raw denomination and get the ibc_denom
-		ibcDenom := transfertypes.ParseDenomTrace(prefixedDenom).IBCDenom()
-		ts, err := k.leverage.GetTokenSettings(ctx, ibcDenom)
+		ibcDenom := ics20types.ParseDenomTrace(prefixedDenom).IBCDenom()
+		ts, err := k.leverage.GetTokenSettings(*k.ctx, ibcDenom)
 		if err != nil {
-			// skip if token is not a registered token on leverage
 			if ltypes.ErrNotRegisteredToken.Is(err) {
-				return nil
+				return nil // skip recording inflow if the token is not registered
 			}
+			k.ctx.Logger().Error("can't get x/leverage token settings", "error", err)
+			return channeltypes.NewErrorAcknowledgement(err)
 		}
 
 		// get the exchange price (eg: UMEE) in USD from oracle using SYMBOL Denom eg: `UMEE`
@@ -257,12 +257,13 @@ func (k Keeper) RecordIBCInflow(ctx sdk.Context,
 		}
 		// calculate total exchange rate
 		powerReduction := ten.Power(uint64(ts.Exponent))
-		inflowInUSD := sdk.MustNewDecFromStr(dataAmount).Quo(powerReduction).Mul(exchangeRate)
+		inflowInUSD := sdk.MustNewDecFromStr(amount).Quo(powerReduction).Mul(exchangeRate)
 
-		tokenInflow := sdk.NewDecCoinFromDec(ibcDenom, inflowInUSD)
+		tokenInflow := k.GetTokenInflow(ibcDenom)
+		tokenInflow.Amount = tokenInflow.Amount.Add(inflowInUSD)
 		k.SetTokenInflow(tokenInflow)
-		totalInflowSum := k.GetTotalInflow()
-		k.SetTotalInflow(totalInflowSum.Add(inflowInUSD))
+		totalInflowSum := k.GetInflowSum()
+		k.SetInflowSum(totalInflowSum.Add(inflowInUSD))
 	}
 
 	return nil