+++ x/uibc/uics20/ics4_wrapper_test.go
@@ -1,9 +1,7 @@
-package quota_test
+package uics20_test
 
 import (
 	"testing"
-
-	"gotest.tools/v3/assert"
 
 	"github.com/cosmos/cosmos-sdk/codec"
 	storetypes "github.com/cosmos/cosmos-sdk/store/types"
@@ -14,6 +12,7 @@ import (
 	clienttypes "github.com/cosmos/ibc-go/v7/modules/core/02-client/types"
 	porttypes "github.com/cosmos/ibc-go/v7/modules/core/05-port/types"
 	"github.com/golang/mock/gomock"
+	"gotest.tools/v3/assert"
 
 	"github.com/umee-network/umee/v6/tests/tsdk"
 	lfixtures "github.com/umee-network/umee/v6/x/leverage/fixtures"
@@ -23,7 +22,7 @@ import (
 	"github.com/umee-network/umee/v6/x/uibc"
 	"github.com/umee-network/umee/v6/x/uibc/mocks"
 	"github.com/umee-network/umee/v6/x/uibc/quota"
-	"github.com/umee-network/umee/v6/x/uibc/quota/keeper"
+	"github.com/umee-network/umee/v6/x/uibc/uics20"
 )
 
 type MockICS4Wrapper struct {
@@ -53,7 +52,7 @@ func TestSendPacket(t *testing.T) {
 
 	storeKey := storetypes.NewMemoryStoreKey("quota")
 	ctx, _ := tsdk.NewCtxOneStore(t, storeKey)
-	kb := keeper.NewKeeperBuilder(codec.NewProtoCodec(nil), storeKey, leverageMock, oracleMock, eg)
+	kb := quota.NewKeeperBuilder(codec.NewProtoCodec(nil), storeKey, leverageMock, oracleMock, eg)
 	dp := uibc.DefaultParams()
 	keeper := kb.Keeper(&ctx)
 	keeper.SetParams(dp)
@@ -62,8 +61,7 @@ func TestSendPacket(t *testing.T) {
 	leverageMock.EXPECT().GetTokenSettings(ctx, "umee").Return(ltypes.Token{}, ltypes.ErrNotRegisteredToken).AnyTimes()
 	oracleMock.EXPECT().Price(ctx, "TEST").Return(sdk.Dec{}, types.ErrMalformedLatestAvgPrice)
 
-	// quota ics5
-	ics4 := quota.NewICS4(mock, kb)
+	ics4 := uics20.NewICS4(mock, kb)
 
 	// error test cases
 	_, err := ics4.SendPacket(ctx, nil, "", "", clienttypes.NewHeight(1, 1), 1, nil)